Move FriendGroup and FriendGroups to contact.friendgroup
