+++ mirai-core-api/src/commonMain/kotlin/Bot.kt
@@ -15,7 +15,7 @@ package net.mamoe.mirai
 import kotlinx.coroutines.*
 import me.him188.kotlin.jvm.blocking.bridge.JvmBlockingBridge
 import net.mamoe.mirai.contact.*
-import net.mamoe.mirai.data.FriendGroups
+import net.mamoe.mirai.contact.friendgroup.FriendGroups
 import net.mamoe.mirai.event.EventChannel
 import net.mamoe.mirai.event.events.BotEvent
 import net.mamoe.mirai.message.action.BotNudge
