+++ mirai-core-api/src/commonMain/kotlin/contact/friendgroup/FriendGroups.kt
@@ -7,7 +7,7 @@
  * https://github.com/mamoe/mirai/blob/dev/LICENSE
  */
 
-package net.mamoe.mirai.data
+package net.mamoe.mirai.contact.friendgroup
 
 import me.him188.kotlin.jvm.blocking.bridge.JvmBlockingBridge
