+++ mirai-core/src/commonMain/kotlin/contact/FriendGroupImpl.kt
@@ -10,7 +10,7 @@
 package net.mamoe.mirai.internal.contact
 
 import net.mamoe.mirai.contact.Friend
-import net.mamoe.mirai.data.FriendGroup
+import net.mamoe.mirai.contact.friendgroup.FriendGroup
 import net.mamoe.mirai.internal.QQAndroidBot
 import net.mamoe.mirai.internal.contact.info.FriendGroupInfo
 import net.mamoe.mirai.internal.network.protocol.packet.list.FriendList
